movie details work in progress

+++ app/src/main/java/com/blueprint/architecture/myapplication/network/APIInterface.kt
@@ -16,5 +16,5 @@ interface APIInterface {
     fun getPopularMovies(@Query("page") pageNumber : Int): Call<PopularMoviesResponse?>?
 
     @GET("/movie/{movie_id}")
-    fun getMovieDetails(@Path("movie_id") movieId : Long) : Call<MovieDetails>
+    fun getMovieDetails(@Path("movie_id") movieId : Int) : Call<MovieDetails>
 }
